requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
folium>=0.14.0
streamlit-folium>=0.13.0
psycopg2-binary>=2.9.0

firebase-admin>=6.2.0
scikit-learn>=1.3.0
prophet>=1.1.4
openai>=1.0.0
requests>=2.31.0
reportlab>=4.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
streamlit-chat>=0.1.1
streamlit-autorefresh>=0.0.6
websockets>=11.0.0
Pillow>=10.0.0
numpy>=1.24.0